requirements.txt
Flask
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
requests
python-dotenv
gunicorn
psycopg2-binary
WTForms
flask-socketio
sqlalchemy
eventlet
flask-cors
email-validator
pyotp
flask-wtf  # Required for FlaskForm in forms.py
flask-bootstrap  # Optional, if Bootstrap is used in templates
flask-mail  # Required for email functionality (e.g., password reset)
flask-limiter
redis
flask-limiter[redis]
